Optimized parts
It was reported that the retainer was over constrained, so i made to tolerances for the M3 screws looser to allow for some play in case people don't align their inserts perfectly. It's recommended that users buy heat set inserts with guides to keep them aligned, but with the new tolerances it should be more forgiving with generic inserts.

I also Improved the tolerances of the M5 screw holes as per request, this however has only been done to the set screw(m3) version. 
I will look into updating the Plunger(m6) version too at a later date.